requirements.txt
langchain==0.1.16
langchain_community==0.0.33
langchain_openai==0.0.8
langchain-groq
mlflow==2.12.1
faiss-cpu 
starlette
fastapi
watchfiles
slowapi
gunicorn
waitress
load_dotenv
dagshub
transformers
torch 
torchvision
sentence-transformers 
scikit-learn 
rouge-score
google-genai
